refactor: Separando o tratamento de erros da bibliteca Joi em um outro arquivo para manter o código do controller mais limpo
